requirements.txt
azure-functions==1.7.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pyodbc==4.0.30
requests==2.25.1
urllib3==1.26.4
